[MISC][RL] Move component to its own folder
